% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{apply_attr_types}
\alias{apply_attr_types}
\title{apply_attr_types}
\usage{
apply_attr_types(x, vr)
}
\arguments{
\item{x}{Vector. Variable data.}

\item{vr}{Character. Relevant variable.}
}
\value{
Replacement vectors.
}
\description{
apply_attr_types
}
