% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{apply_var_types}
\alias{apply_var_types}
\title{apply_var_types}
\usage{
apply_var_types(x, vr)
}
\arguments{
\item{x}{Vector. Variable data.}

\item{vr}{Character. Relevant variable: Unified variable name.}
}
\value{
Replacement vector.
}
\description{
apply_var_types
}
