% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{sdsdecoding}
\alias{sdsdecoding}
\alias{lookup_everything}
\alias{lookup_vars}
\alias{lookup_var_complete_names}
\alias{lookup_var_types}
\alias{apply_var_types}
\alias{lookup_attrs}
\alias{lookup_attr_types}
\alias{apply_attr_types}
\alias{lookup_IGerM_category}
\title{sdsanalysis decoding functions}
\usage{
lookup_everything(sds_df)

lookup_vars(var_ids)

lookup_var_complete_names(var_short_names)

lookup_var_types(var_short_names)

apply_var_types(var_data, var_short_name)

lookup_attrs(var_data, var_short_name)

lookup_attr_types(var_data, var_short_name)

apply_attr_types(var_data, var_short_name)

lookup_IGerM_category(igerm_data, subcategory = FALSE)
}
\arguments{
\item{sds_df}{Dataframe. Data.frame in SDS standard format.}

\item{var_ids}{Character Vector. Variable IDs.}

\item{var_short_names}{Character Vector. Variable short names.}

\item{var_data}{Vector. Variable data.}

\item{var_short_name}{Character. Variable short name.}

\item{igerm_data}{Character vector. IGerM character codes in data.}

\item{subcategory}{Boolean. Should the function return IGerM subcategories 
instead of categories?}
}
\description{
SDS traditionally provides a set of predefined values for each variable. 
That's not just convenience: It theoretically also allows for a high degree of 
comparability between different datasets. This predefined values/categories are 
encoded with a simple and minimalistic alphanumerical scheme. That's a 
technological rudiment both from the time when the systems that served SDS as 
an inspiration were created and when most stone tool analysis was made without 
a computer in reach.

The encoding has the big disadvantage that it's not immediately human readable.
If you try to understand an SDS dataset you're forced to constantly look up new 
variables in the SDS publications. That makes it very difficult to get a fast 
overview of an SDS dataset.

sdsanalysis offers functions to quickly decode the cryptic codes in the SDS 
tables and replace them with human readable descriptions. This is implemented 
with hash tables to enable high-speed transformation even for datasets with 
thousands of artefacts.

\itemize{
  \item{\code{\link{lookup_everything}}: Wizard function. Enter a SDS data.frame and 
  receive a decoded version}
  \item{\code{\link{lookup_vars}}: Get short variable names from IDs}
  \item{\code{\link{lookup_var_complete_names}}: Get long variable names from short 
  variable names}
  \item{\code{\link{lookup_var_types}}: Get variable data types from short variable 
  names}
  \item{\code{\link{apply_var_types}}: Get variable vector with correct data type from 
  variable vector with arbitrary data type}
  \item{\code{\link{lookup_attrs}}: Get decoded version of encoded variable vector}
  \item{\code{\link{lookup_attr_types}}: Get variable vector with semantic type from 
  variable vector}
  \item{\code{\link{apply_attr_types}}: Get variable vector with the correct values set 
  to NA based on the semantic type vector from variable vector}
  \item{\code{\link{lookup_IGerM_category}}: Get IGerM category or subcategory vector 
  from IGerM vector}
}
}
