% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdsdownload.R
\name{sdsdownload}
\alias{sdsdownload}
\alias{get_single_artefact_data}
\alias{get_multi_artefact_data}
\alias{get_description}
\alias{get_site}
\alias{get_coords}
\alias{get_dating}
\alias{get_creator}
\alias{get_type_options}
\alias{get_available_datasets}
\alias{get_all_sds_data_urls}
\alias{get_variables_list}
\alias{get_variable_values_list}
\title{sdsanalysis data download functions}
\usage{
get_single_artefact_data(dataset_names)

get_multi_artefact_data(dataset_names)

get_description(dataset_names)

get_site(dataset_names)

get_coords(dataset_names)

get_dating(dataset_names)

get_creator(dataset_names)

get_type_options(dataset_names)

get_available_datasets()

get_all_sds_data_urls()

get_variables_list()

get_variable_values_list()
}
\arguments{
\item{dataset_names}{Character vector. Names of available datasets.}
}
\description{
sdsanalysis offers the following functions to access available SDS data.

\itemize{
  \item{\code{\link{get_available_datasets}}: Get a list of datasets that can 
  be directly downloaded with sdsanalysis}
  \item{\code{\link{get_type_options}}: Get the types of data that are available 
  for one/multiple datasets (single - \emph{Einzelaufnahme}, multi - \emph{Sammelaufnahme})}
  \item{\code{\link{get_single_artefact_data}} / 
  \code{\link{get_multi_artefact_data}}: Download one/multiple datasets as a 
  dataframe/a list of dataframes}
  \item{\code{\link{get_description}}: Download description text of one/
  multiple datasets}
  \item{\code{\link{get_site}}: Get site names of one/multiple datasets}
  \item{\code{\link{get_coords}}: Get site coordinates of one/multiple datasets}
  \item{\code{\link{get_dating}}: Get period information of one/multiple datasets}
  \item{\code{\link{get_creator}}: Get author of one/multiple datasets}
}
}
