% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_plots.R
\name{dynamic_plot}
\alias{dynamic_plot}
\title{dynamic_plot}
\usage{
dynamic_plot(inputdata, var1, var2)
}
\arguments{
\item{inputdata}{test}

\item{var1}{test}

\item{var2}{test}
}
\value{
ggplot object.
}
\description{
dynamic_plot
}
