% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_metadata.R
\name{get_coords}
\alias{get_coords}
\title{get_coords}
\usage{
get_coords(dataset_names)
}
\arguments{
\item{dataset_names}{Character vector. Names of available datasets.}
}
\description{
get_coords
}
