% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_metadata.R
\name{get_creator}
\alias{get_creator}
\title{get_creator}
\usage{
get_creator(dataset_names)
}
\arguments{
\item{dataset_names}{Character vector. Names of available datasets.}
}
\description{
get_creator
}
