% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_metadata.R
\name{get_dating}
\alias{get_dating}
\title{get_dating}
\usage{
get_dating(dataset_names)
}
\arguments{
\item{dataset_names}{Character vector. Names of available datasets.}
}
\description{
get_dating
}
