% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_description}
\alias{get_description}
\title{get_description}
\usage{
get_description(dataset_name)
}
\arguments{
\item{dataset_name}{Character. Name of an available dataset.}
}
\description{
get_description
}
