% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_metadata}
\alias{get_metadata}
\title{lookup_data_positions}
\usage{
get_metadata(dataset_name, type)
}
\arguments{
\item{dataset_name}{Character. Name of an available dataset.}
}
\description{
lookup_data_positions
}
