% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_multi_artefact_data}
\alias{get_multi_artefact_data}
\title{get_multi_artefact_data}
\usage{
get_multi_artefact_data(dataset_names)
}
\arguments{
\item{dataset_names}{Character vector. Names of available datasets.}
}
\description{
get_multi_artefact_data
}
