% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_metadata.R
\name{get_type_options}
\alias{get_type_options}
\title{get_type_options}
\usage{
get_type_options(dataset_names)
}
\arguments{
\item{dataset_names}{Character vector. Names of available datasets.}
}
\description{
get_type_options
}
