% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_attr_types}
\alias{lookup_attr_types}
\title{lookup_attr_types}
\usage{
lookup_attr_types(x, vr)
}
\arguments{
\item{x}{Vector. Keys to look up values.}

\item{vr}{Character. Relevant variable.}
}
\value{
Variable names.
}
\description{
lookup_attr_types
}
