% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_attrs}
\alias{lookup_attrs}
\title{lookup_attrs}
\usage{
lookup_attrs(x, vr)
}
\arguments{
\item{x}{Vector. Keys to look up values: Numeric or character values in database.}

\item{vr}{Character. Relevant variable: Unified variable name.}
}
\value{
Transformed variable values.
}
\description{
lookup_attrs
}
