% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{lookup_data_positions}
\alias{lookup_data_positions}
\title{lookup_data_positions}
\usage{
lookup_data_positions(dataset_name)
}
\arguments{
\item{dataset_name}{Character. Name of an available dataset.}
}
\description{
lookup_data_positions
}
