% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_everything}
\alias{lookup_everything}
\title{lookup_everything}
\usage{
lookup_everything(x, fb)
}
\arguments{
\item{x}{Dataframe. Table in SDS standard format.}

\item{fb}{Numeric. Relevant Formblatt.}
}
\value{
Decoded and optimized SDS data.frame.
}
\description{
lookup_everything
}
