% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_IGerM_category}
\alias{lookup_IGerM_category}
\title{lookup_IGerM_category}
\usage{
lookup_IGerM_category(x, subcategory = FALSE)
}
\arguments{
\item{x}{Character vector. Keys to look up values: IGerM character codes in data.}

\item{subcategory}{Boolean. Should the function return IGerM subcategories instead of categories?}
}
\description{
lookup_IGerM_category
}
