% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_var_types}
\alias{lookup_var_types}
\title{lookup_var_types}
\usage{
lookup_var_types(vr)
}
\arguments{
\item{vr}{Character. Keys to look up values: Unified variable names.}
}
\value{
Variable types (names of types as character vector).
}
\description{
lookup_var_types
}
