% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdsdecode.R
\name{sdsdecoding}
\alias{sdsdecoding}
\alias{lookup_everything}
\alias{lookup_vars}
\alias{lookup_var_complete_names}
\alias{lookup_var_types}
\alias{apply_var_types}
\alias{lookup_attrs}
\alias{lookup_attr_types}
\alias{apply_attr_types}
\alias{lookup_IGerM_category}
\title{sdsanalysis decoding functions}
\usage{
lookup_everything(sds_df)

lookup_vars(var_ids)

lookup_var_complete_names(var_short_names)

lookup_var_types(var_short_names)

apply_var_types(var_data, var_short_name)

lookup_attrs(var_data, var_short_name)

lookup_attr_types(var_data, var_short_name)

apply_attr_types(var_data, var_short_name)

lookup_IGerM_category(igerm_data, subcategory = FALSE)
}
\arguments{
\item{sds_df}{Dataframe. Data.frame in SDS standard format.}

\item{var_ids}{Character Vector. Variable IDs.}

\item{var_short_names}{Character Vector. Variable short names.}

\item{var_data}{Vector. Variable data.}

\item{var_short_name}{Character. Variable short name.}

\item{igerm_data}{Character vector. IGerM character codes in data.}

\item{subcategory}{Boolean. Should the function return IGerM subcategories 
instead of categories?}
}
\description{
SDS traditionally provides a set of predefined values for each variable. 
That's not just convenience: It theoretically also allows for a high degree of 
comparability between different datasets. This predefined values/categories are 
encoded with a simple and minimalistic alphanumerical scheme. That's a 
technological rudiment both from the time when the systems that served SDS as 
an inspiration were created and when most stone tool analysis was made without 
a computer in reach.

The encoding has the big disadvantage that it's not immediately human readable.
If you try to understand a SDS dataset you're forced to constantly look up new 
variables in the 
\href{https://github.com/Johanna-Mestorf-Academy/sdsbrowser#references}{SDS publications}. 
That makes it very difficult to get a fast overview.

sdsanalysis offers functions to quickly decode the cryptic codes in the SDS 
tables and replace them with human readable descriptions. This is implemented 
with hash tables to enable high-speed transformation even for datasets with 
thousands of artefacts. The hash tables are compiled from two reference tables for 
\href{https://github.com/Johanna-Mestorf-Academy/sdsanalysis/blob/master/data-raw/variable_list.csv}{variables}
and 
\href{https://github.com/Johanna-Mestorf-Academy/sdsanalysis/blob/master/data-raw/variable_values_list.csv}{variable values}. 

\itemize{
  \item{\code{\link{lookup_everything}}: Wizard function. Enter a 
  \href{https://github.com/Johanna-Mestorf-Academy/sdsbrowser#a-new-dataset}{SDS data.frame}
  and receive a decoded version. This function employs the ones below and some more
  helpers to make the decoding process as simple as possible}
  \item{\code{\link{lookup_vars}}: 
    \strong{In:} character vector with variable IDs (e.g. \emph{FB1_23, FB2_56})
    \strong{Out:} character vector with short variable names (\emph{menge_rinde, dorsal_praep})
  }
  \item{\code{\link{lookup_var_complete_names}}: 
    \strong{In:} character vector with short variable names (\emph{menge_rinde, dorsal_praep})
    \strong{Out:} character vector with long variable names  
    (e.g. \emph{Art der Dorsalflaechenpraeparation, Menge der Rinde und natuerlichen Sprungflaeche}) 
  }
  \item{\code{\link{lookup_var_types}}: 
    \strong{In:} character vector with short variable names (\emph{menge_rinde, dorsal_praep})
    \strong{Out:} character vector with variable data types (e.g. \emph{character, numeric})
  }
  \item{\code{\link{apply_var_types}}: 
    \strong{In:} encoded variable vector (SDS data.frame column) + respective variable short name
    \strong{Out:} encoded variable vector with corrected data type
  } 
  \emph{apply} instead of \emph{lookup}, because in this case the result of an other lookup is 
  used to manipulate the input vector.
  \item{\code{\link{lookup_attrs}}: 
    \strong{In:} variable vector (SDS data.frame column) + respective variable short name
    \strong{Out:} decoded variable vector
  }
  \item{\code{\link{lookup_attr_types}}: 
    \strong{In:} encoded variable vector (SDS data.frame column) + respective variable short name
    \strong{Out:} character vector with semantic type (e.g. \emph{normal, unknown})
  }
  \item{\code{\link{apply_attr_types}}: 
    \strong{In:} encoded variable vector (SDS data.frame column) + respective variable short name
    \strong{Out:} encoded variable vector with the correct values set to NA based on the semantic 
    type
  }
  \item{\code{\link{lookup_IGerM_category}}: 
    \strong{In:} decoded IGerM vector
    \strong{Out:} IGerM category or subcategory vector 
  }
}
}
