% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_description}
\alias{get_description}
\title{get_description}
\usage{
get_description(dataset_names)
}
\arguments{
\item{dataset_names}{Character vector. Names of available datasets.}
}
\description{
get_description
}
