% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_attr_types}
\alias{lookup_attr_types}
\title{lookup_attr_types}
\usage{
lookup_attr_types(x, vr)
}
\arguments{
\item{x}{Vector. Keys to look up values: Numeric or character values in database.}

\item{vr}{Character. Relevant variable: Unified variable name.}
}
\value{
Type of variable: Not data type, but "normal"/"absent"/"unknown".
}
\description{
lookup_attr_types
}
