% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_attrs}
\alias{lookup_attrs}
\title{lookup_attrs}
\usage{
lookup_attrs(x, vr)
}
\arguments{
\item{x}{Character Vector. Keys to look up values.}

\item{vr}{Character. Relevant variable.}
}
\value{
Variable names.
}
\description{
lookup_attrs
}
