% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_var_complete_names}
\alias{lookup_var_complete_names}
\title{lookup_var_complete_names}
\usage{
lookup_var_complete_names(vr)
}
\arguments{
\item{vr}{Character Vector. Short variable names to look up.}
}
\value{
Complete variable names.
}
\description{
lookup_var_complete_names
}
