% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_var_types}
\alias{lookup_var_types}
\title{lookup_var_types}
\usage{
lookup_var_types(x, vr)
}
\arguments{
\item{x}{Vector. Variable data.}

\item{vr}{Character. Relevant variable.}
}
\value{
Variable names.
}
\description{
lookup_var_types
}
