% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_vars}
\alias{lookup_vars}
\title{lookup_vars}
\usage{
lookup_vars(x)
}
\arguments{
\item{x}{Character Vector. Keys to look up values: Variable IDs.}
}
\value{
Unified variable names.
}
\description{
lookup_vars
}
